requirements.txt
openai==1.59.6
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.28.0
pdf2image==1.16.3
Pillow==10.4.0
